project/requirements.txt
Flask==3.1.1
gunicorn==23.0.0
numpy==1.26.4
Pillow==11.2.1
tensorflow-cpu==2.15.1
h5py==3.10.0
scikit-learn==1.5.1